fix(config): tighten payout_lag_days to uint8
Mirrors the Rust u8 on PeriodConfig (period.rs). Rejects out-of-range payout_lag_days at Go YAML unmarshal instead of truncating at the Rust boundary. The rules.go > 30 warning compares fine under uint8. Updates the type-aware assertion in TestUnmarshalMinimalUnilevelPlan. Adds boundary tests (255 ok; 256/300/-1 rejected).

+++ internal/config/types.go
@@ -31,7 +31,7 @@ type CompensationPlan struct {
 type PeriodConfig struct {
 	Length        string  `yaml:"length" json:"length"`
 	StartDate     *string `yaml:"start_date" json:"start_date"`
-	PayoutLagDays int     `yaml:"payout_lag_days" json:"payout_lag_days"`
+	PayoutLagDays uint8   `yaml:"payout_lag_days" json:"payout_lag_days"`
 }
 
 // --- Volume ---
